feat: [GP-63] WEB - Implement Waiting for approval screen
+ PrivateRoute: add waiting-for-approval to welcome session
+ Change url img to public/assets/images
+ Delete unnecessary lib
+ Edit responsive after add header
